Rename jhealpix_reference -> jhealpix
